fix: update export parameters file and transform into a function
