+++ src/CMakeLists.txt
@@ -4,8 +4,5 @@ rock_library(genset_whisperpower_ddc
     DEPS_PKGCONFIG base-types iodrivers_base)
 
 rock_executable(genset_whisperpower_ddc_ctl Main.cpp
-    DEPS genset_whisperpower_ddc)
-
-rock_executable(export_parameters ExportParameters.cpp
     DEPS genset_whisperpower_ddc
     DEPS_PKGCONFIG jsoncpp)
